+++ tensorflow_decision_forests/keras/wrappers_pre_generated.py
@@ -805,6 +805,7 @@ class GradientBoostedTreesModel(core.CoreModel):
   decision trees trained sequentially. Each tree is trained to predict and then
   "correct" for the errors of the previously trained trees (more precisely each
   tree predict the gradient of the loss relative to the model output).
+  GBTs use [early stopping](early_stopping.md) to avoid overfitting.
 
   Usage example:
 
@@ -1682,7 +1683,7 @@ class RandomForestModel(core.CoreModel):
   dataset (sampled with replacement).
   
   The algorithm is unique in that it is robust to overfitting, even in extreme
-  cases e.g. when there is more features than training examples.
+  cases e.g. when there are more features than training examples.
   
   It is probably the most well-known of the Decision Forest training
   algorithms.